+++ python/lsst/ts/idl/enums/LaserTracker.py
@@ -11,39 +11,20 @@
 # the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.
 #
-# This program is distributed in the hope that it will be useful,
+# This program is distributed in the hope that it will be useful
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
 #
 # You should have received a copy of the GNU General Public License
 
-from . import ATCamera
-from . import ATDome
-from . import ATHexapod
-from . import ATMCS
-from . import ATMonochromator
-from . import ATPneumatics
-from . import ATPtg
-from . import ATSpectrograph
-from . import ATThermoelectricCooler
-from . import ATWhiteLight
-from . import Electrometer
-from . import FiberSpectrograph
-from . import Guider
-from . import LaserTracker
-from . import LinearStage
-from . import MTAOS
-from . import MTDome
-from . import MTHexapod
-from . import MTM1M3
-from . import MTM2
-from . import MTMount
-from . import MTPtg
-from . import MTRotator
-from . import PMD
-from . import Scheduler
-from . import Script
-from . import ScriptQueue
-from . import TunableLaser
-from . import Watcher
+__all__ = ["SalIndex"]
+
+import enum
+
+
+class SalIndex(enum.IntEnum):
+    """Allowed SAL indices."""
+
+    MTAlignment = 1
+    OTHER = 2